+++ mini-project/src/main/java/com/fast/miniproject/auth/service/Impl/UserServiceImpl.java
@@ -5,17 +5,16 @@ import com.fast.miniproject.auth.entity.User;
 import com.fast.miniproject.auth.jwt.JwtProvider;
 import com.fast.miniproject.auth.repository.UserRepository;
 import com.fast.miniproject.auth.service.UserService;
+import com.fast.miniproject.global.config.RedisService;
 import com.fast.miniproject.global.response.ErrorResponseDTO;
 import com.fast.miniproject.global.response.ResponseDTO;
 import com.fast.miniproject.product.service.ProductService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.NoSuchElementException;
-import java.util.concurrent.TimeUnit;
 
 @Service
 @RequiredArgsConstructor
