--- mini-project/src/main/java/com/fast/miniproject/auth/service/RefreshTokenService.java
@@ -1,9 +1,0 @@
-package com.fast.miniproject.auth.service;
-
-import com.fast.miniproject.global.response.ResponseDTO;
-
-public interface RefreshTokenService {
-
-    public ResponseDTO<?> validateRefreshToken(String refreshToken);
-
-}
