+++ mini-project/src/main/java/com/fast/miniproject/auth/service/TokenService.java
@@ -5,5 +5,6 @@ import com.fast.miniproject.global.response.ResponseDTO;
 public interface TokenService {
 
     ResponseDTO<?> logout(String header);
+    public ResponseDTO<?> validateRefreshToken(String refreshToken);
     boolean checkLogout(String token);
 }
